feat(audio): float FFT magnitudes, with band energy averaged in the linear domain
getByteFrequencyData quantises to 256 steps across [minDecibels, maxDecibels]
before we ever see the data, which throws away precision exactly where it is
scarcest — quiet bands near the floor. Switch to getFloatFrequencyData.

Not a swap: the two differ in DOMAIN, not just precision. The byte path handed
us a linear-in-dB ramp already mapped to 0..255; the float path hands us raw
dBFS. Averaging dB values directly computes a geometric mean of amplitudes,
which is not a band level and under-reads any band containing a peak. So each
bin is now converted to power (10^(dB/10)), averaged, returned to dB, and only
then mapped to 0..1 by the new `dbToUnit`.

SCALE IS DELIBERATELY UNCHANGED. `dbToUnit` applies the same [dbFloor,
dbCeiling] → [0,1] ramp getByteFrequencyData applied internally, so
NORMALIZE_SILENCE_FLOOR, NORMALIZE_MIN_PEAK, AGC_FLOOR, every rule's
thresholdMin/Max and the meter's amber/red points all keep their calibration —
nothing was retuned. The filterbank's own follower and AGC tests pass
unchanged, which is the evidence for that claim; a new test pins the mapping at
floor/mid/ceiling and asserts a flat spectrum reads back its exact input level.

Also note getFloatFrequencyData IGNORES the analyser's min/maxDecibels — those
only ever affected byte conversion. The dB window is now ours to apply, which
is why it lives in one place.

Per-bin power is computed once per frame into a reused scratch buffer rather
than inside the band loop: bands already share bins, and will share more once
windowed kernels land.

+++ engine/features/modulation/ModulationEngine.ts
@@ -261,7 +261,7 @@ class ModulationEngine {
      *  for a harder hit. Users trim from there with the rule's Gain knob. */
     private static readonly TRANSIENT_FULL_SCALE = 20;
 
-    private processAudioSignal(rule: ModulationRule, _data: Uint8Array, delta: number): number {
+    private processAudioSignal(rule: ModulationRule, _data: Float32Array, delta: number): number {
         if (rule.freqEnd <= rule.freqStart) return 0;
 
         // Read the fractional-octave BANDS, not raw bins. The bands are already
